training/requirements.txt
pyyaml
pydash
jsgf2fst==0.1.2
jsonlines
doit
